Refactor UI Views
Moves macOS-specific UI components into dedicated files for better organization and maintainability. This change improves code structure and readability.
